+++ sdks/rust/tests/case-conversion-client/src/module_bindings/add_person_2_type.rs
@@ -8,7 +8,7 @@ use spacetimedb_sdk::__codegen::{self as __sdk, __lib, __sats, __ws};
 #[sats(crate = __lib)]
 pub struct AddPerson2 {
     pub first_3_name: String,
-    pub player_ref: u64,
+    pub player_ref: u32,
     pub age_value: u8,
     pub score_total: u32,
 }
