+++ sdks/rust/tests/case-conversion-client/src/module_bindings/ban_player_1_type.rs
@@ -7,8 +7,8 @@ use spacetimedb_sdk::__codegen::{self as __sdk, __lib, __sats, __ws};
 #[derive(__lib::ser::Serialize, __lib::de::Deserialize, Clone, PartialEq, Debug)]
 #[sats(crate = __lib)]
 pub struct BanPlayer1 {
-    pub player_1_id: u64,
-    pub ban_until_6: u64,
+    pub player_1_id: u32,
+    pub ban_until_6: u32,
 }
 
 impl __sdk::InModule for BanPlayer1 {
