+++ sdks/rust/tests/case-conversion-client/src/module_bindings/player_2_status_type.rs
@@ -9,7 +9,7 @@ use spacetimedb_sdk::__codegen::{self as __sdk, __lib, __sats, __ws};
 pub enum Player2Status {
     Active1,
 
-    BannedUntil(u64),
+    BannedUntil(u32),
 }
 
 impl __sdk::InModule for Player2Status {
